ContributionReward : check params validity (#572)
* ContributionReward : check params validity

* Fix genericScheme execute function

* uint to uint256

require scheme register before getting its param

* FixReputationAllocation - indexed _beneficiary  BeneficiaryAddressAdded evet

+++ contracts/Migrations.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.4.25;
 
 contract Migrations {
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     modifier restricted() {
         if (msg.sender == owner) {
@@ -15,7 +15,7 @@ contract Migrations {
         owner = msg.sender;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         lastCompletedMigration = completed;
     }
