+++ contracts/test/ARCDebug.sol
@@ -15,7 +15,7 @@ contract ARCDebug {
     event LogAddress(address _msg);
     event LogInt(int _msg);
     event LogString(string _msg);
-    event LogUint(uint _msg);
+    event LogUint(uint256 _msg);
     event LogBytes(bytes _msg);
     event LogBytes32(bytes32 _msg);
     event LogBool(bool _msg);
