+++ contracts/test/ExternalTokenLockerMock.sol
@@ -6,7 +6,7 @@ contract ExternalTokenLockerMock {
     // user => amount
     mapping (address => uint) public lockedTokenBalances;
 
-    function lock(uint _amount) public {
+    function lock(uint256 _amount) public {
         lockedTokenBalances[msg.sender] = _amount;
     }
 }
